scraper/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
firecrawl-py
pydantic
google-genai
exa-py
PyPDF2
pytest==7.4.3
pytest-mock==3.12.0
linkedin-scraper
redis==5.0.1